requirements.txt
fastapi[standard]
pyjwt
pwdlib[argon2]
streamlit
pandas
numpy
opencv-python
tensorflow
pillow
torch
transformers
nltk
